refactor(cli): relocate trace merger
